Refactor: Change logo files for new style
