+++ src/ui/Footer/Footer.module.css
@@ -68,8 +68,7 @@ footer {
 }
 
 .logo {
-  width: 120px;
-  filter: grayscale(100%);
+  width: 100px;
 }
 
 @media only screen and (max-width: 700px) {
