+++ src/ui/HeadBar/HeadBar.module.css
@@ -41,7 +41,7 @@
 }
 
 .logo {
-  height: 50px;
+  height: 35px;
   width: auto;
   user-select: none;
   -webkit-user-drag: none;
